Refactoring: renamed getJobsUseCase to getTestResultsUseCase
